Prta03_09まで完了

+++ Practice/src/practice02/PTra02_01.java
@@ -15,9 +15,10 @@ waru=20/5;
 System.out.println(waru);
 
 		// ★ 126に13を掛けた値を出力してください
-int kake;
+long kake;
 kake=126*13;
 System.out.println(kake);
+
 		// ★ 97を6で割ったときの余りを出力してください
 int amari;
 amari=97%6;
